From a GUI, lauch via emacsclient by default
* etc/emacs-mail.desktop, etc/emacs.desktop: Call emacsclient not
  emacs by default.  Provide plain emacs as an alternative action.

Fix desktop launcher

The New Window action is redundant because it is now catered for by
the default action.

* etc/emacs.desktop: Ensure we pass --create-frame to emacsclient if
we lack at least one filename.

Document behaviour of emacs.desktop

* doc/emacs/misc.texi (Using Emacs as a Server): Explain that
emacs.desktop will start a daemon if emacsclient can’t connect.

When launching from .desktop, open new frame if there is none

* etc/emacs.desktop, etc/emacs-mail.desktop: Specify --display so
that, in daemon mode, emacs knows where to open a frame if it can’t
find an existing one to use.